FactorRotations

Documentation for FactorRotations.

FactorRotations.CrawfordFerguson

FactorRotations.IterationState

FactorRotations.MinimumEntropy

FactorRotations.Oblimin

FactorRotations.Oblique

FactorRotations.Orthogonal

FactorRotations.Quartimax

FactorRotations.RotationMethod

FactorRotations.RotationState

FactorRotations.RotationType

FactorRotations._rotate

FactorRotations.criterion_and_gradient

FactorRotations.initialize

FactorRotations.isconverged

FactorRotations.project!

FactorRotations.rotate

LinearAlgebra.rotate!

FactorRotations.CrawfordFerguson — Type

CrawfordFerguson(; kappa, orthogonal = false)

The family of Crawford-Ferguson rotation methods.

Keyword arguments

kappa: T

Details

The Crawford-Ferguson family allows both orthogonal and oblique rotation of the p-byk-factor loading matrix. If orthogonal rotation is performed, Crawford-Ferguson is equivalent to Oblimin rotation given the following values for kappa:

kappa = 0: Quartimax

kappa = 1/p: Varimax

kappa = k/2p: Equamax

kappa = (k - 1)/(p + k - 2): Parsimax

kappa = 1: Factor parsimony

FactorRotations.IterationState — Type

IterationState

A struct that holds the state of an iteration of the rotation algorithm.

FactorRotations.MinimumEntropy — Type

minetropy

FactorRotations.Oblimin — Type

Oblimin(; gamma, orthogonal = false)

The family of Oblimin rotation methods.

Keyword arguments

gamma: The shape parameter determining the rotation criterion (see Details).

orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false`)

Details

The Oblimin rotation family allow orthogonal as well as oblique rotation of the factor loading matrix. If orthogonal rotation is performed, Oblimin is equivalent to the following rotation methods given a value for gamma:

gamma = 0: Quartimax

gamma = 0.5: Biquartimax

gamma = 1: Varimax

gamma = p/2: Equamax (where p is the number of columns in the factor loading matrix)

For oblique rotation Oblimin is equivalent to the following rotation methods:

gamma = 0: Quartimin

gamma = 0.5: Biquartimin

Examples

julia> Oblimin(gamma = 0.5)
Oblimin{Oblique, Float64}(0.5)

julia> Oblimin(gamma = 1, orthogonal = true)
Oblimin{Orthogonal, Int64}(1)

FactorRotations.Oblique — Type

Oblique

A type representing an oblique rotation type.

FactorRotations.Orthogonal — Type

Orthogonal

A type representing an orthogonal rotation type.

FactorRotations.Quartimax — Type

Quartimax

FactorRotations.RotationMethod — Type

RotationMethod{T<:RotationType}

An abstract type representing a factor rotation method.

Each implementation of M <: RotationMethod must implement criterion_and_gradient.

FactorRotations.RotationState — Type

RotationState

A struct that holds the state of the rotation algorithm.

Fields

init: The initial rotation matrix

A: The initial factor loading matrix

T: The current state of the rotation matrix

L: The current state of the rotated loading matrix

iterations: A vector of IterationState that holds iteration states of the optimization.

FactorRotations.RotationType — Type

RotationType

An abstract type representing a type of factor rotation.

FactorRotations._rotate — Method

_rotate(A::AbstractMatrix, method::RotationMethod{Orthogonal}; kwargs...)

Implements the algorithm for factor rotation described in Bernaard & Jennrich (2005) for orthogonal factor rotation.

FactorRotations.criterion_and_gradient — Function

criterion_and_gradient(method::RotationMethod, Λ::AbstractMatrix{<:Real})

Calculate the criterion and gradient of a given method with respect to the factor loading matrix Λ.

Returns a Tuple with the criterion value as the first element and gradient as the second element.

FactorRotations.initialize — Method

initialize(init, A)

Initialize a RotationState with initial values init and original loading matrix A. If init = nothing, the identity matrix will be used as initial values.

FactorRotations.isconverged — Method

isconverged(s, atol)

determines the convergence status of the algorithm.

FactorRotations.project! — Method

project!(Gp, T, G)

Compute the projection Gp of G for orthogonal matrices.

FactorRotations.rotate — Method

rotate(Λ, method::RotationMethod; kwargs...)
rotate!(Λ, method::RotationMethod; kwargs...)

Perform a rotation of the factor loading matrix Λ.

Keyword arguments

atol: Sets the absolute tolerance for convergence of the algorithm (default: 1e-6).

alpha: Sets the inital value for alpha (default: 1).

maxiter1: Controls the number of maximum iterations in the outer loop of the algorithm (default: 1000).

maxiter2: Controls the number of maximum iterations in the inner loop of the algorithm (default: 10).

init: A k-by-k matrix of starting values for the algorithm. If init = nothing (the default), the identity matrix will be used as starting values.

LinearAlgebra.rotate! — Method

rotate(Λ, method::RotationMethod; kwargs...)

Perform a rotation of the factor loading matrix Λ and overwrite Λ with the rotated loading matrix.

See also rotate.