API

Rotation criteria

FactorRotations.Absolmin — Type

Absolmin(epsilon = 0)

The Absolmin component loss factor rotation criterion. It has the loss function

\[h(\lambda) = |\lambda|\]

FactorRotations.Biquartimax — Function

Biquartimax()

The Biquartimax rotation method.

Details

The Biquartimax rotation method is a special case of the Oblimin rotation with parameters gamma = 0.5 and orthogonal = true.

Examples

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> L_biquartimax = rotate(L, Biquartimax());

julia> L_oblimin = rotate(L, Oblimin(gamma = 0.5, orthogonal = true));

julia> loadings(L_biquartimax) ≈ loadings(L_oblimin)
true

FactorRotations.ComponentLoss — Type

ComponentLoss(loss::Function; orthogonal = false)

A generic implementation of the component loss factor rotation method. loss defines the loss function that is applied to the components of the loading matrix.

Keyword arguments

orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false)

Details

The component loss factor rotation applies a loss function to each element of the factor loading matrix. Then the following criterion is minimized:

\[Q(\Lambda) = \sum_i \sum_j h(\lambda_{ij})\]

Examples

Quartimax as a component loss

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> quartimax_loss = ComponentLoss(x -> x^4, orthogonal = true);

julia> L_component_loss = rotate(L, quartimax_loss)
FactorRotation{Float64} with loading matrix:
8×2 Matrix{Float64}:
 0.898755  0.194824
 0.933943  0.129749
 0.902131  0.103864
 0.876508  0.171284
 0.315572  0.876476
 0.251123  0.773489
 0.198007  0.714678
 0.307857  0.659334

julia> L_quartimax = rotate(L, Quartimax())
FactorRotation{Float64} with loading matrix:
8×2 Matrix{Float64}:
 0.898755  0.194823
 0.933943  0.129748
 0.902132  0.103864
 0.876508  0.171284
 0.315572  0.876476
 0.251124  0.773489
 0.198008  0.714678
 0.307858  0.659334

julia> isapprox(loadings(L_component_loss), loadings(L_quartimax), atol = 1e-5)
true

FactorRotations.Concave — Type

Concave(bandwidth = 1)

The simple concave component loss factor rotation criterion. It has the loss function

\[h(\lambda) = 1 - \exp(-\frac{|\lambda|}{b})\]

where $b$ is the bandwidth parameter.

FactorRotations.CrawfordFerguson — Type

CrawfordFerguson(; kappa, orthogonal = false)

The family of Crawford-Ferguson rotation methods.

Keyword arguments

kappa: The parameter determining the rotation criterion (see Details).

orthogonal: orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false)

Details

The Crawford-Ferguson family allows both orthogonal and oblique rotation of the p-byk-factor loading matrix. If orthogonal rotation is performed, Crawford-Ferguson is equivalent to Oblimin rotation given the following values for kappa:

kappa = 0 → Quartimax

kappa = 1/p → Varimax

kappa = k/2p → Equamax

kappa = (k - 1)/(p + k - 2) → Parsimax

kappa = 1 → Factor parsimony

Examples

julia> CrawfordFerguson(kappa = 0, orthogonal = true)
CrawfordFerguson{Orthogonal, Int64}(0)

julia> CrawfordFerguson(kappa = 1/2, orthogonal = false)
CrawfordFerguson{Oblique, Float64}(0.5)

FactorRotations.Infomax — Type

Infomax(; orthogonal = false)

The Infomax rotation method.

Keyword arguments

orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false)

FactorRotations.KatzRohlf — Type

KatzRohlf(bandwidth)

A component loss criterion with loss function

\[h(\lambda) = 1 - \exp(-(\frac{\lambda}{b})^2)\]

where $b$ is the bandwidth parameter.

FactorRotations.LinearRightConstant — Type

LinearRightConstant(bandwidth)

The linear right constant component loss factor rotation criterion. It has the loss function

\[h(\lambda) = \begin{cases} (\frac{\lambda}{b})^2&\text{if } |\lambda| \leq b \\ 1 &\text{if } \lambda > b \end{cases}\]

where $b$ is the bandwidth parameter.

FactorRotations.MinimumEntropy — Type

MinimumEntropy()

The Minimum Entropy rotation method.

FactorRotations.MinimumEntropyRatio — Type

MinimumEntropyRatio()

The Minimum Entropy Ratio rotation method.

FactorRotations.Oblimax — Type

Oblimax(; orthogonal = false)

The Oblimax rotation method.

Keyword arguments

orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false)

Details

The Oblimax rotation method is equivalent to Quartimax for orthogonal rotation.

Examples

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> L_oblimax = rotate(L, Oblimax(orthogonal = true));

julia> L_quartimax = rotate(L, Quartimax());

julia> isapprox(loadings(L_oblimax), loadings(L_quartimax), atol = 1e-6)
true

FactorRotations.Oblimin — Type

Oblimin(; gamma, orthogonal = false)

The family of Oblimin rotation methods.

Keyword arguments

gamma: The shape parameter determining the rotation criterion (see Details).

orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false)

Details

The Oblimin rotation family allow orthogonal as well as oblique rotation of the factor loading matrix. If orthogonal rotation is performed, Oblimin is equivalent to the following rotation methods given a value for gamma:

gamma = 0 → Quartimax

gamma = 0.5 → Biquartimax

gamma = 1 → Varimax

gamma = p/2 → Equamax

For oblique rotation Oblimin is equivalent to the following rotation methods:

gamma = 0 → Quartimin

gamma = 0.5 → Biquartimin

Examples

julia> Oblimin(gamma = 0.5)
Oblimin{Oblique, Float64}(0.5)

julia> Oblimin(gamma = 1, orthogonal = true)
Oblimin{Orthogonal, Int64}(1)

FactorRotations.PatternSimplicity — Type

PatternSimplicity(; orthogonal = false)

The Pattern Simplicity factor rotation criterion.

Keyword arguments

orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false)

FactorRotations.Quartimax — Type

Quartimax()

The Quartimax rotation criterion.

Details

The Quartimax criterion is a special case of the Oblimin rotation criterion with parameter gamma = 0.

Examples

Setting up the criterion

julia> Quartimax()
Quartimax()

Testing equivalence of Quartimax and Oblimin

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> L_quartimax = rotate(L, Quartimax());

julia> L_oblimin = rotate(L, Oblimin(gamma = 0, orthogonal = true));

julia> loadings(L_quartimax) ≈ loadings(L_oblimin)
true

FactorRotations.TandemCriteria — Type

TandemCriteria(; keep)

The tandem criteria rotation method.

Keyword arguments

keep: The number of factors to keep for the second tandem criterion.

FactorRotations.TandemCriterionI — Type

TandemCriterionI()

The first criterion of the tandem criteria factor rotation method.

FactorRotations.TandemCriterionII — Type

TandemCriterionI()

The second criterion of the tandem criteria factor rotation method.

FactorRotations.TargetRotation — Type

TargetRotation(target::AbstractMatrix; orthogonal = false)

The (partial) target rotation criterion.

Keyword arguments

orthogonal: If orthogonal = true an orthogonal rotation is performed, an oblique rotation otherwise. (default: false)

Details

Target rotation rotates a factor loading matrix towards the target matrix, target. For a fully specified target matrix (e.g. all entries in the matrix are numbers), full target rotation is performed.

Partially specified target rotation can be achieved setting the unspecified entries in the target matrix to missing.

Examples

Full target rotation

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> target = [1 0; 1 0; 1 0; 1 0; 0 1; 0 1; 0 1; 0 1];

julia> rotate(L, TargetRotation(target, orthogonal = true))
FactorRotation{Float64} with loading matrix:
8×2 Matrix{Float64}:
 0.882633  0.258215
 0.922358  0.195806
 0.892467  0.167726
 0.862116  0.233154
 0.252473  0.89669
 0.195508  0.789382
 0.146707  0.726945
 0.260213  0.679549

Partially specified target rotation

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> target = [1 0; missing missing; 1 0; 1 0; 0 1; 0 1; 0 1; 0 1];

julia> rotate(L, TargetRotation(target, orthogonal = true))
FactorRotation{Float64} with loading matrix:
8×2 Matrix{Float64}:
 0.873299  0.288209
 0.915133  0.227193
 0.886218  0.198109
 0.85365   0.262462
 0.221701  0.90479
 0.168434  0.795599
 0.121793  0.731532
 0.236852  0.68804

FactorRotations.Varimax — Type

Varimax()

The Varimax rotation criterion.

Details

The Varimax criterion is a special case of the Oblimin rotation criterion with parameter gamma = 1.

Examples

Setting up the criterion

julia> Varimax()
Varimax()

Testing equivalence of Varimax and Oblimin

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> L_varimax = rotate(L, Varimax());

julia> L_oblimin = rotate(L, Oblimin(gamma = 1, orthogonal = true));

julia> loadings(L_varimax) ≈ loadings(L_oblimin)
true

User Interface

FactorRotations.setverbosity! — Function

setverbosity!(::Bool)

Sets the global verbosity level of the package. If set to false (the default), package functions will not log @info statements. If set to true, package functions will provide @info statements.

FactorRotations.rotate — Function

rotate(Λ, method::RotationMethod; kwargs...)

Perform a rotation of the factor loading matrix Λ using a rotation method.

Keyword arguments

atol: Sets the absolute tolerance for convergence of the algorithm (default: 1e-6).

alpha: Sets the inital value for alpha (default: 1).

maxiter1: Controls the number of maximum iterations in the outer loop of the algorithm (default: 1000).

maxiter2: Controls the number of maximum iterations in the inner loop of the algorithm (default: 10).

init: A k-by-k matrix of starting values for the algorithm. If init = nothing (the default), the identity matrix will be used as starting values.

verbose: Print logging statements (default: false)

Examples

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> rotate(L, Varimax())
FactorRotation{Float64} with loading matrix:
8×2 Matrix{Float64}:
 0.886061  0.246196
 0.924934  0.183253
 0.894664  0.155581
 0.865205  0.221416
 0.264636  0.893176
 0.206218  0.786653
 0.156572  0.724884
 0.269424  0.67595

LinearAlgebra.rotate! — Function

rotate!(Λ, method::RotationMethod; kwargs...)

Perform a rotation of the factor loading matrix Λ and overwrite Λ with the rotated loading matrix.

For a list of available keyword arguments see rotate.

Examples

julia> L = [
           0.830 -0.396
           0.818 -0.469
           0.777 -0.470
           0.798 -0.401
           0.786  0.500
           0.672  0.458
           0.594  0.444
           0.647  0.333
       ];

julia> rotate!(L, Quartimax())
8×2 Matrix{Float64}:
 0.898755  0.194823
 0.933943  0.129748
 0.902132  0.103864
 0.876508  0.171284
 0.315572  0.876476
 0.251124  0.773489
 0.198008  0.714678
 0.307858  0.659334

FactorRotations.isorthogonal — Function

isorthogonal(::RotationMethod)

Checks if the supplied rotation method is orthogonal.

Examples

julia> isorthogonal(Varimax())
true

julia> isorthogonal(Oblimax(orthogonal = false))
false

FactorRotations.isoblique — Function

isoblique(::RotationMethod)

Checks if the supplied rotation method is oblique.

Examples

julia> isoblique(Varimax())
false

julia> isoblique(Oblimax(orthogonal = false))
true

FactorRotations.loadings — Function

loadings(r::FactorRotation)

Return the rotated factor loading matrix from r.

FactorRotations.rotation — Function

rotation(r::FactorRotation)

Return the factor rotation matrix from r.

FactorRotations.factor_correlation — Function

factor_correlation(r::FactorRotation)

Return the factor correlation matrix from r.

Internals

FactorRotations.FactorRotation — Type

FactorRotation{T <: Real}

A type holding results of a factor rotation.

Fields

L: The rotated factor loading matrix

T: The factor rotation matrix

phi: The factor correlation matrix

FactorRotations.RotationMethod — Type

RotationMethod{T<:RotationType}

An abstract type representing a factor rotation method.

Each implementation of M <: RotationMethod must implement at least one of the following methods:

criterion

criterion_and_gradient

If criterion is implemented, gradients are calculated by automatic differentiation.

FactorRotations.Orthogonal — Type

Orthogonal

A type representing an orthogonal rotation type.

FactorRotations.Oblique — Type

Oblique

A type representing an oblique rotation type.

FactorRotations.criterion — Function

criterion(method::RotationMethod, Λ::Abstractmatrix{<:Real})

Calculate the criterion of a given method with respect to the factor loading matrix Λ.

FactorRotations.criterion_and_gradient — Function

criterion_and_gradient(method::RotationMethod, Λ::AbstractMatrix{<:Real})

Calculate the criterion and gradient of a given method with respect to the factor loading matrix Λ.

Returns a Tuple with the criterion value as the first element and gradient as the second element.